Removes throwing of exception
To integrate with the APIRequest object exceptions are no longer thrown for invalid IDs. Instead, all invalid values are given the value of -1. This enables the model to return false when a query is run for item ID=-1. The APIRequest object will then return page not found instead of an empty data object.