Adds additional http header return types and other tweaks.
Minor tweak to createLinks so that the element 'self' now returns the start and limit values. This is used to help inform pagination within the APIRequest object.

getAll() now includes resultsPerPage element which indicates the expected number of items. This is used for pagination, as currentResults value will change on the final page.