Refactors object to include more data for getOne and getALl.
getOne now includes a boolean field indicating whether the athlete won the fight or not. This is used to display badges for winners/when there is a draw.

getAll now includes data for the event headline fight, e.g. names and athlete image.

Removed the option to filter for upcoming events/past events. The implementation of this is overcomplicated and unnecessary.