Fixes a bug with outcome badges showing as 'draw' when winnerId is null
winnerId is null when there is no winner - this could mean the fight has not taken place yet, or there was a draw. However, there was no way to tell the difference previously.

The fix addresses this by distinguishing outcomes, when fight outcome is null this is reflective of the fight not having taken place. Using this, winnerId is set to -1. This means that no badge will be shown.

Also updates placeholder text to TBC when no outcome is recorded.