Implements functionality to add and update fight results
Due to the normalisation of the data, to add/update fight result requires that fight athlete data is also updated. Both FightResult and FightAthlete models have been updated to support this.

Given that FightResults are updated via FightID, it was necessary to update the CRUDController so that it used a different different pathway when requesting FightAthlete data (default implementation of getOne function uses the module ID).